% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cristin-other.R
\name{nva_cristin_funding_source}
\alias{nva_cristin_funding_source}
\title{Get a specific funding source}
\usage{
nva_cristin_funding_source(id)
}
\arguments{
\item{id}{Funding source identifier}
}
\value{
A list with funding source details
}
\description{
Retrieves information about a specific funding source by ID.
}
\examples{
\dontrun{
source <- nva_cristin_funding_source("NFR")
}
}
