% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cristin-other.R
\name{nva_cristin_keyword}
\alias{nva_cristin_keyword}
\title{Get a Cristin keyword by ID}
\usage{
nva_cristin_keyword(keyword_id)
}
\arguments{
\item{keyword_id}{Keyword identifier}
}
\value{
A list with the keyword record
}
\description{
Retrieves a specific keyword from the Cristin registry.
}
\examples{
\dontrun{
keyword <- nva_cristin_keyword("12345")
}
}
