% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cristin-person.R
\name{nva_cristin_person}
\alias{nva_cristin_person}
\title{Get a Cristin person by ID}
\usage{
nva_cristin_person(cristin_id)
}
\arguments{
\item{cristin_id}{Cristin person identifier (numeric or character)}
}
\value{
A list containing the person record:
\describe{
\item{id}{Cristin person ID}
\item{names}{List with firstName and lastName}
\item{affiliations}{List of organizational affiliations}
\item{identifiers}{External identifiers (ORCID, etc.)}
}
}
\description{
Retrieves information about a researcher from the Cristin registry.
}
\examples{
\dontrun{
# Get a researcher by Cristin ID
person <- nva_cristin_person(12345)

# Access their name
paste(person$names$firstName, person$names$lastName)
}
}
