% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nva-package.R
\docType{package}
\name{nva-package}
\alias{nva}
\alias{nva-package}
\title{nva: R Client for the Norwegian National Research Archive (NVA) API}
\description{
Provides functions for accessing the Norwegian National Research Archive (NVA) API. Includes support for searching publications, retrieving metadata, accessing Cristin data, and downloading files.
}
\author{
\strong{Maintainer}: Roar Bakken Stovner \email{robast@oslomet.no}

}
\keyword{internal}
